#1 feat : Layout page divide two section & result page
